pc/scripts/requirements.txt
pyserial
influxdb-client
